backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Knot integration
requests==2.32.3
python-dotenv==1.0.1

# AI agents
dedalus-labs
pdfplumber==0.10.3


